第四次提交，添加DS18B20温度传感器驱动。并可在实际开发板上运行led、数码管、DS18B20驱动。protues DS18B20未知。

+++ test_led/DEVICES/SEG/seg.h
@@ -8,7 +8,7 @@
 #define ST_CP 	PAout(7)	// PA7
 
 void segInit(void);	//初始化		
-void seg4Display(float temp); //seg显示
+void seg4Display(int temp); //seg显示
 void seg4Test(void);//seg测试函数
 void segOut(unsigned char X); //seg 8位输出函数
